Moving some classes to the Data subfolder.
