+++ Aegis.Pumps.Tests/AegisServiceClientTests.cs
@@ -679,7 +679,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Aegis.Core;
+using Aegis.Core.Data;
 using Jil;
 
 namespace Aegis.Pumps.Tests
