+++ Aegis.Pumps/Actions.cs
@@ -679,6 +679,7 @@ using System;
 using System.Net;
 using System.Net.Http.Headers;
 using Aegis.Core;
+using Aegis.Core.Data;
 
 namespace Aegis.Pumps
 {
