+++ Aegis.Pumps/BlackListClient.cs
@@ -679,6 +679,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using Aegis.Core;
+using Aegis.Core.Data;
 
 namespace Aegis.Pumps
 {
