+++ Aegis.Pumps/SchedulerJobs/SendAegisEventsJob.cs
@@ -679,6 +679,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Aegis.Core;
+using Aegis.Core.Data;
 
 namespace Aegis.Pumps.SchedulerJobs
 {
